+++ Theresa3rd-Bot/TheresaBot.Main/Handler/LoliconHandler.cs
@@ -67,7 +67,7 @@ namespace TheresaBot.Main.Handler
                 workMsgs.Add(new PlainContent(loliconBusiness.getWorkInfo(loliconData, todayLeftCount, template)));
 
                 PixivSetuContent setuContent = new PixivSetuContent(workMsgs, setuFiles, loliconData);
-                int[] msgIds = await command.ReplyGroupSetuAsync(setuContent, BotConfig.SetuConfig.RevokeInterval, BotConfig.PixivConfig.SendImgBehind, true);
+                long[] msgIds = await command.ReplyGroupSetuAsync(setuContent, BotConfig.SetuConfig.RevokeInterval, BotConfig.PixivConfig.SendImgBehind, true);
                 Task recordTask = recordBusiness.AddPixivRecord(setuContent, msgIds, command.GroupId);
                 if (BotConfig.SetuConfig.SendPrivate)
                 {
