+++ Theresa3rd-Bot/TheresaBot.Main/Handler/LolisukiHandler.cs
@@ -70,7 +70,7 @@ namespace TheresaBot.Main.Handler
                 workMsgs.Add(new PlainContent(lolisukiBusiness.getWorkInfo(lolisukiData, startDateTime, todayLeftCount, template)));
 
                 PixivSetuContent setuContent = new PixivSetuContent(workMsgs, setuFiles, lolisukiData);
-                int[] msgIds = await command.ReplyGroupSetuAsync(setuContent, BotConfig.SetuConfig.RevokeInterval, BotConfig.PixivConfig.SendImgBehind, true);
+                long[] msgIds = await command.ReplyGroupSetuAsync(setuContent, BotConfig.SetuConfig.RevokeInterval, BotConfig.PixivConfig.SendImgBehind, true);
                 Task recordTask = recordBusiness.AddPixivRecord(setuContent, msgIds, command.GroupId);
                 if (BotConfig.SetuConfig.SendPrivate)
                 {
