+++ Theresa3rd-Bot/TheresaBot.Main/Handler/PixivSubscribeHandler.cs
@@ -405,7 +405,7 @@ namespace TheresaBot.Main.Handler
                 workMsgs.Add(new PlainContent($"pixiv画师[{pixivWorkInfo.userName}]的最新作品："));
                 workMsgs.Add(new PlainContent(pixivBusiness.getWorkInfo(pixivWorkInfo, BotConfig.PixivConfig.Template)));
                 PixivSetuContent setuContent = new PixivSetuContent(workMsgs, setuFiles, pixivWorkInfo);
-                int[] msgIds = await Session.SendGroupMessageAsync(command.GroupId, setuContent, BotConfig.PixivConfig.SendImgBehind);
+                long[] msgIds = await Session.SendGroupMessageAsync(command.GroupId, setuContent, BotConfig.PixivConfig.SendImgBehind);
                 Task recordTask = recordBusiness.AddPixivRecord(setuContent, msgIds, command.GroupId);
             }
             catch (Exception ex)
