+++ Theresa3rd-Bot/TheresaBot.Main/Helper/SessionHelper.cs
@@ -5,27 +5,27 @@ namespace TheresaBot.Main.Helper
 {
     public static class SessionHelper
     {
-        public static async Task<int[]> SendGroupMessageAsync(this BaseSession session, long groupId, SetuContent setuContent, bool sendImgBehind)
+        public static async Task<long[]> SendGroupMessageAsync(this BaseSession session, long groupId, SetuContent setuContent, bool sendImgBehind)
         {
-            if (setuContent is null) return new int[0];
+            if (setuContent is null) return new long[0];
             List<BaseContent> setuInfos = setuContent.SetuInfos ?? new();
             List<FileInfo> setuFiles = setuContent.SetuImages ?? new();
             if (sendImgBehind)
             {
-                int workMsgId = await session.SendGroupMessageAsync(groupId, setuInfos);
+                long workMsgId = await session.SendGroupMessageAsync(groupId, setuInfos);
                 await Task.Delay(1000);
-                int imgMsgId = await session.SendGroupMessageAsync(groupId, setuFiles.ToBaseContent());
-                return new int[] { workMsgId, imgMsgId };
+                long imgMsgId = await session.SendGroupMessageAsync(groupId, setuFiles.ToBaseContent());
+                return new long[] { workMsgId, imgMsgId };
             }
             else
             {
                 List<BaseContent> msgList = setuInfos.Concat(setuFiles.ToBaseContent()).ToList();
-                int msgId = await session.SendGroupMessageAsync(groupId, msgList.ToArray());
-                return new int[] { msgId };
+                long msgId = await session.SendGroupMessageAsync(groupId, msgList.ToArray());
+                return new long[] { msgId };
             }
         }
 
-        public static async Task<int> SendGroupMergeAsync(this BaseSession session, long groupId, List<SetuContent> setuContents)
+        public static async Task<long> SendGroupMergeAsync(this BaseSession session, long groupId, List<SetuContent> setuContents)
         {
             if (setuContents.Count == 0) return 0;
             return await session.SendGroupMergeAsync(groupId, setuContents.ToBaseContents());
