+++ Theresa3rd-Bot/TheresaBot.Main/Relay/BaseRelay.cs
@@ -2,13 +2,13 @@
 {
     public abstract class BaseRelay
     {
-        public int MsgId { get; set; }
+        public long MsgId { get; set; }
 
         public string Message { get; set; }
 
         public long MemberId { get; init; }
 
-        public BaseRelay(int msgId, string message, long memberId)
+        public BaseRelay(long msgId, string message, long memberId)
         {
             this.MsgId = msgId;
             this.Message = message;
