+++ Theresa3rd-Bot/TheresaBot.Main/Relay/GroupRelay.cs
@@ -2,9 +2,9 @@
 {
     public abstract class GroupRelay : BaseRelay
     {
-        public int GroupId { get; set; }
+        public long GroupId { get; set; }
 
-        public GroupRelay(int msgId, string message, long groupId, long memberId) : base(msgId, message, memberId)
+        public GroupRelay(long msgId, string message, long groupId, long memberId) : base(msgId, message, memberId)
         {
             this.GroupId = GroupId;
         }
